Development - new ui and new skills on upgrade queue rework (#17)
* CSR-1 - changing unity version to 6000.4.5f1

* CSR-1 - changing unity version to 6000.4.5f1 (#1)

* SCR-2 - add agent md with technology, project, bestpractices docs and refactor / documenting skills

* CSR-2 - documenting current systems

* CSR-2 - update folder structure in skills

* CSR-3 - update main pr pipeline

* CSR-3 - fix unity version in pipelines

* Csr 4 general refactoring (#4)

* Initial project setup (#3)

* CSR-1 - changing unity version to 6000.4.5f1

* CSR-1 - changing unity version to 6000.4.5f1 (#1)

* SCR-2 - add agent md with technology, project, bestpractices docs and refactor / documenting skills

* CSR-2 - documenting current systems

* CSR-2 - update folder structure in skills

* CSR-3 - update main pr pipeline

* Change Unity Pipelines Version

commit a85a75757748db48722daa3bbc8bd3efb29e9107
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Tue May 5 19:00:27 2026 +0200

    CSR-3 - fix unity version in pipelines

commit 8a5e61429b33b64e48fc99f836aca82f77a2baa3
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Tue May 5 18:31:34 2026 +0200

    CSR-3 - update main pr pipeline

commit bbfdeeedf1d8ca26e3bf08b63dcd80867132ae7c
Merge: f55370b cb6e90e
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Tue May 5 18:28:13 2026 +0200

    Merge branch 'CSR-2-ai-native-initial' into development

commit f55370bfcc13b39dffc60b200d9f5da346633651
Merge: 94f3751 2e3d293
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Tue May 5 18:27:51 2026 +0200

    Merge branch 'CSR-1-changing-unity-version-6000-4-5f1' into development

commit cb6e90e0fbe438245d42aca4c8c8b8bcaa0c6a1a
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Sun May 3 23:39:34 2026 +0200

    CSR-2 - update folder structure in skills

commit 56077c0b5af4cae40398c72c585e1723225f1e31
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Sun May 3 23:33:28 2026 +0200

    CSR-2 - documenting current systems

commit 3f6e716c8723f71fe13681dce40bfdc2f90c3824
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Sun May 3 23:13:16 2026 +0200

    SCR-2 - add agent md with technology, project, bestpractices docs and refactor / documenting skills

commit 94f37512c3d92a191bb8272654de3a7b0e786785
Author: Konrad Blaszyński <77151114+XenkQ@users.noreply.github.com>
Date:   Sun May 3 22:42:20 2026 +0200

    CSR-1 - changing unity version to 6000.4.5f1 (#1)

commit 2e3d293baa75d18473706d9c4006fc14f33ca40b
Author: Konrad Blaszyński <blaszynski.konrad@gmail.com>
Date:   Sun May 3 22:40:29 2026 +0200

    CSR-1 - changing unity version to 6000.4.5f1

* CSR-4 - code refactoring

* CSR-4 - scripts folder restructurization

* CSR-4 - fix rotation

* CSR-5 - add minimal in game skill upgrade popup (#5)

* CSR-2 - reorganizing angets informations folder structure to context (#6)

* Csr 6 grid and flow field optimalizations (#8)

* CSR-6 - creation of navigation performance optimalization plan

* CSR-6 - main camera fix

* CSR-6 - injecting camera instead of camera main

* CSR-6 - navigation optimalizations

* CSR-8 - fixed skills bugs and introduced new optimalizations for them + quick menu disable fix (#9)

* CSR 9 - new rare enemies with separate spawning system (#10)

* CSR-9 - add new barrel enemy

* CSR-9 - exp particles fix

* CSR-10 - new UI (#11)

* CSR-10 - damage numbers visibility change

* CSR-10 - main menu new ui

* CSR-10 - fix main menu buttons sound and game objects bindings

* CSR-10 - new settings menu

* CSR-10 - settings dropdown and storage path changes

* CSR-10 - gitignore update

* CSR-10 - new exp bar

* CSR-10 - new death screen

* CSR-10 - upgrade presenter new ui

* CSR-10 - fixing skills upgrade button audio issue

* Csr 10 new UI (#12)

* CSR-10 - damage numbers visibility change

* CSR-10 - main menu new ui

* CSR-10 - fix main menu buttons sound and game objects bindings

* CSR-10 - new settings menu

* CSR-10 - settings dropdown and storage path changes

* CSR-10 - gitignore update

* CSR-10 - new exp bar

* CSR-10 - new death screen

* CSR-10 - upgrade presenter new ui

* CSR-10 - fixing skills upgrade button audio issue

* CSR-10 - add new pause and timer ui

* CSR-10 - in game cursor fix

* CSR-10 - new score board ui

* CSR-10 - buttons pack on skill upgrade and skill selection + new lighting on new skill menu

* Csr 11 skill acquire and update rework (#14)

* CSR-11 - acquire skills after levels interval and toogle unlimited stats

* CSR-11 - changing how skills are currently upgraded (now new skill after 3 upgrades).

* CSR-11 - fixing skills unlimited upgrades error caused by byte type

* Csr 12 car drifting (#15)

* CSR-12 - add drifting

* CSR-12 - added basic drifting

* Csr 10 new UI (decorations added) (#16)

* CSR-10 - damage numbers visibility change

* CSR-10 - main menu new ui

* CSR-10 - fix main menu buttons sound and game objects bindings

* CSR-10 - new settings menu

* CSR-10 - settings dropdown and storage path changes

* CSR-10 - gitignore update

* CSR-10 - new exp bar

* CSR-10 - new death screen

* CSR-10 - upgrade presenter new ui

* CSR-10 - fixing skills upgrade button audio issue

* CSR-10 - add new pause and timer ui

* CSR-10 - in game cursor fix

* CSR-10 - new score board ui

* CSR-10 - buttons pack on skill upgrade and skill selection + new lighting on new skill menu

* CSR-10 - adding decoration to score board and death menu

+++ Assets/Scripts/Skills/SkillsRegistry.cs
@@ -8,7 +8,7 @@ namespace Assets.Scripts.Skills
     public interface ISkillsRegistry
     {
         public IReadOnlyList<ISkillBase> Skills { get; }
-        public byte UninitializedSkillsCount { get; }
+        public int UninitializedSkillsCount { get; }
 
         public IEnumerable<ISkillBase> GetUninitializedSkills();
 
@@ -18,7 +18,7 @@ namespace Assets.Scripts.Skills
     public class SkillsRegistry : MonoBehaviour, ISkillsRegistry
     {
         public IReadOnlyList<ISkillBase> Skills { get; private set; }
-        public byte UninitializedSkillsCount { get; private set; }
+        public int UninitializedSkillsCount { get; private set; }
 
         private void Awake()
         {
@@ -27,7 +27,7 @@ namespace Assets.Scripts.Skills
 
         private void Start()
         {
-            UninitializedSkillsCount = (byte)GetUninitializedSkills().Count();
+            UninitializedSkillsCount = GetUninitializedSkills().Count();
 
             InitializeSkill(Skills[0]);
         }
